Browsing departments (#3)
* Emit CustomerBrowsing event to MQTT
* Rename storeLocationService to customerStateService
* Add sequenceCnt and produce browsing event base on that
* Add logging
* Add mnv wrapper
* Add Quarkus Multistage Docker Build (#2)
* Add missing endlines
* Restore surfire plugin in pom.xml
* Add envrionment variables description
* Get rid of polish channel name
* Change coordiantes to float (from int)
* Unify env names (with other components); Add default constructor for BrowsingEvent
* Update CustomerBrowsing - @RegisterForReflection
* Update CustomerBrowsing.java - add import

+++ src/main/java/org/demo/rsotf/model/CustomerMovement.java
@@ -2,8 +2,8 @@ package org.demo.rsotf.model;
 
 public class CustomerMovement {
     private String id;
-    private int x;
-    private int y;
+    private double x;
+    private double y;
     private int ts;
 
     private String seenIn;
@@ -23,11 +23,11 @@ public class CustomerMovement {
         return id;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
@@ -35,11 +35,11 @@ public class CustomerMovement {
         this.id = id;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
